remove the ability for code to change file permissions, this was only needed for pluginmanager's test, and pluginmanager doesn't even run with securitymanager yet.
