+++ core/src/main/resources/org/elasticsearch/bootstrap/security.policy
@@ -88,9 +88,6 @@ grant {
   // otherwise can be provided only to test libraries
   permission java.lang.RuntimePermission "fileSystemProvider";
 
-  // needed by plugin manager to set unix permissions
-  permission java.lang.RuntimePermission "accessUserInformation";
-
   // needed by jvminfo for monitoring the jvm
   permission java.lang.management.ManagementPermission "monitor";
