+++ core/src/test/java/org/elasticsearch/plugins/PluginInfoTests.java
@@ -35,22 +35,9 @@ import static org.hamcrest.Matchers.contains;
 
 public class PluginInfoTests extends ESTestCase {
 
-    static void writeProperties(Path pluginDir, String... stringProps) throws IOException {
-        assert stringProps.length % 2 == 0;
-        Files.createDirectories(pluginDir);
-        Path propertiesFile = pluginDir.resolve(PluginInfo.ES_PLUGIN_PROPERTIES);
-        Properties properties =  new Properties();
-        for (int i = 0; i < stringProps.length; i += 2) {
-            properties.put(stringProps[i], stringProps[i + 1]);
-        }
-        try (OutputStream out = Files.newOutputStream(propertiesFile)) {
-            properties.store(out, "");
-        }
-    }
-
     public void testReadFromProperties() throws Exception {
         Path pluginDir = createTempDir().resolve("fake-plugin");
-        writeProperties(pluginDir,
+        PluginTestUtil.writeProperties(pluginDir,
             "description", "fake desc",
             "name", "my_plugin",
             "version", "1.0",
@@ -71,7 +58,7 @@ public class PluginInfoTests extends ESTestCase {
 
     public void testReadFromPropertiesNameMissing() throws Exception {
         Path pluginDir = createTempDir().resolve("fake-plugin");
-        writeProperties(pluginDir);
+        PluginTestUtil.writeProperties(pluginDir);
         try {
             PluginInfo.readFromProperties(pluginDir);
             fail("expected missing name exception");
@@ -79,7 +66,7 @@ public class PluginInfoTests extends ESTestCase {
             assertTrue(e.getMessage().contains("Property [name] is missing in"));
         }
 
-        writeProperties(pluginDir, "name", "");
+        PluginTestUtil.writeProperties(pluginDir, "name", "");
         try {
             PluginInfo.readFromProperties(pluginDir);
             fail("expected missing name exception");
@@ -90,7 +77,7 @@ public class PluginInfoTests extends ESTestCase {
 
     public void testReadFromPropertiesDescriptionMissing() throws Exception {
         Path pluginDir = createTempDir().resolve("fake-plugin");
-        writeProperties(pluginDir, "name", "fake-plugin");
+        PluginTestUtil.writeProperties(pluginDir, "name", "fake-plugin");
         try {
             PluginInfo.readFromProperties(pluginDir);
             fail("expected missing description exception");
@@ -101,7 +88,7 @@ public class PluginInfoTests extends ESTestCase {
 
     public void testReadFromPropertiesVersionMissing() throws Exception {
         Path pluginDir = createTempDir().resolve("fake-plugin");
-        writeProperties(pluginDir, "description", "fake desc", "name", "fake-plugin");
+        PluginTestUtil.writeProperties(pluginDir, "description", "fake desc", "name", "fake-plugin");
         try {
             PluginInfo.readFromProperties(pluginDir);
             fail("expected missing version exception");
@@ -112,7 +99,7 @@ public class PluginInfoTests extends ESTestCase {
 
     public void testReadFromPropertiesJvmAndSiteMissing() throws Exception {
         Path pluginDir = createTempDir().resolve("fake-plugin");
-        writeProperties(pluginDir,
+        PluginTestUtil.writeProperties(pluginDir,
             "description", "fake desc",
             "version", "1.0",
             "name", "my_plugin");
@@ -126,7 +113,7 @@ public class PluginInfoTests extends ESTestCase {
 
     public void testReadFromPropertiesElasticsearchVersionMissing() throws Exception {
         Path pluginDir = createTempDir().resolve("fake-plugin");
-        writeProperties(pluginDir,
+        PluginTestUtil.writeProperties(pluginDir,
             "description", "fake desc",
             "name", "my_plugin",
             "version", "1.0",
@@ -141,7 +128,7 @@ public class PluginInfoTests extends ESTestCase {
 
     public void testReadFromPropertiesJavaVersionMissing() throws Exception {
         Path pluginDir = createTempDir().resolve("fake-plugin");
-        writeProperties(pluginDir,
+        PluginTestUtil.writeProperties(pluginDir,
             "description", "fake desc",
             "name", "my_plugin",
             "elasticsearch.version", Version.CURRENT.toString(),
@@ -158,7 +145,7 @@ public class PluginInfoTests extends ESTestCase {
     public void testReadFromPropertiesJavaVersionIncompatible() throws Exception {
         String pluginName = "fake-plugin";
         Path pluginDir = createTempDir().resolve(pluginName);
-        writeProperties(pluginDir,
+        PluginTestUtil.writeProperties(pluginDir,
             "description", "fake desc",
             "name", pluginName,
             "elasticsearch.version", Version.CURRENT.toString(),
@@ -177,7 +164,7 @@ public class PluginInfoTests extends ESTestCase {
     public void testReadFromPropertiesBadJavaVersionFormat() throws Exception {
         String pluginName = "fake-plugin";
         Path pluginDir = createTempDir().resolve(pluginName);
-        writeProperties(pluginDir,
+        PluginTestUtil.writeProperties(pluginDir,
                 "description", "fake desc",
                 "name", pluginName,
                 "elasticsearch.version", Version.CURRENT.toString(),
@@ -195,7 +182,7 @@ public class PluginInfoTests extends ESTestCase {
 
     public void testReadFromPropertiesBogusElasticsearchVersion() throws Exception {
         Path pluginDir = createTempDir().resolve("fake-plugin");
-        writeProperties(pluginDir,
+        PluginTestUtil.writeProperties(pluginDir,
             "description", "fake desc",
             "version", "1.0",
             "jvm", "true",
@@ -211,7 +198,7 @@ public class PluginInfoTests extends ESTestCase {
 
     public void testReadFromPropertiesOldElasticsearchVersion() throws Exception {
         Path pluginDir = createTempDir().resolve("fake-plugin");
-        writeProperties(pluginDir,
+        PluginTestUtil.writeProperties(pluginDir,
             "description", "fake desc",
             "name", "my_plugin",
             "version", "1.0",
@@ -227,7 +214,7 @@ public class PluginInfoTests extends ESTestCase {
 
     public void testReadFromPropertiesJvmMissingClassname() throws Exception {
         Path pluginDir = createTempDir().resolve("fake-plugin");
-        writeProperties(pluginDir,
+        PluginTestUtil.writeProperties(pluginDir,
             "description", "fake desc",
             "name", "my_plugin",
             "version", "1.0",
@@ -245,7 +232,7 @@ public class PluginInfoTests extends ESTestCase {
     public void testReadFromPropertiesSitePlugin() throws Exception {
         Path pluginDir = createTempDir().resolve("fake-plugin");
         Files.createDirectories(pluginDir.resolve("_site"));
-        writeProperties(pluginDir,
+        PluginTestUtil.writeProperties(pluginDir,
             "description", "fake desc",
             "name", "my_plugin",
             "version", "1.0",
@@ -258,7 +245,7 @@ public class PluginInfoTests extends ESTestCase {
 
     public void testReadFromPropertiesSitePluginWithoutSite() throws Exception {
         Path pluginDir = createTempDir().resolve("fake-plugin");
-        writeProperties(pluginDir,
+        PluginTestUtil.writeProperties(pluginDir,
                 "description", "fake desc",
                 "name", "my_plugin",
                 "version", "1.0",
